Created a modeller package and moved ResidueTemplate there.
