+++ chemistry/modeller/__init__.py
@@ -1,0 +1,11 @@
+"""
+This package contains functionality necessary to carry out basic molecular
+modelling tasks.
+"""
+
+__author__ = 'Jason Swails'
+__date__ = '2015'
+__license__ = 'LGPL'
+
+from chemistry.modeller.residue import ResidueTemplate
+#from chemistry.amber import AmberOFFLibrary
